requirements.txt
slackclient>=2.2.1
certifi